Add common FO page
