+++ tests/puppeteer/pages/FO/home.js
@@ -1,21 +1,15 @@
-const CommonPage = require('../commonPage');
+const FOBasePage = require('../FO/FObasePage');
 
-module.exports = class Home extends CommonPage {
+module.exports = class Home extends FOBasePage {
   constructor(page) {
     super(page);
 
     // Selectors for home page
-    this.logoHomePage = '#_desktop_logo';
-    this.cartProductsCount = '#_desktop_cart span.cart-products-count';
     this.productArticle = '#content .products div:nth-child(%NUMBER) article';
     this.productImg = `${this.productArticle} img`;
     this.productQuickViewLink = `${this.productArticle} a.quick-view`;
-    this.userInfoLink = '#_desktop_user_info';
-    this.logoutLink = `${this.userInfoLink} .user-info a.logout`;
-    this.contactLink = '#contact-link';
     this.allProductLink = '#content a.all-product-link';
     this.totalProducts = '#js-product-list-top .total-products > p';
-    this.categoryMenu = '#category-%ID > a';
     // Quick View modal
     this.quickViewModalDiv = 'div[id*=\'quickview-modal\']';
     this.quantityWantedInput = `${this.quickViewModalDiv} input#quantity_wanted`;
@@ -29,15 +23,7 @@ module.exports = class Home extends CommonPage {
    * Check home page
    */
   async checkHomePage() {
-    await this.page.waitForSelector(this.logoHomePage, {visible: true});
-  }
-
-  /**
-   * go to the home page
-   */
-  async goToHomePage() {
-    await this.waitForSelectorAndClick(this.logoHomePage);
-    this.page.waitForNavigation({waitUntil: 'networkidle0'});
+    await this.page.waitForSelector(this.productImg.replace('%NUMBER', '1'), {visible: true});
   }
 
   /**
@@ -45,45 +31,8 @@ module.exports = class Home extends CommonPage {
    * @param id, product id
    */
   async goToProductPage(id) {
-    await this.page.waitForSelector(this.logoHomePage, {visible: true});
+    await this.page.waitForSelector(this.productImg.replace('%NUMBER', id));
     await this.waitForSelectorAndClick(this.productImg.replace('%NUMBER', id), 5000);
-  }
-
-  /**
-   * Filter by category
-   * @param categoryID, category id from the BO
-   */
-  async filterByCategory(categoryID) {
-    await this.waitForSelectorAndClick(this.categoryMenu.replace('%ID', categoryID));
-  }
-
-  /**
-   * Filter by subcategory
-   * @param categoryID, category id from the BO
-   * @param subCategoryID, subcategory id from the BO
-   */
-  async filterSubCategory(categoryID, subCategoryID) {
-    await this.page.hover(this.categoryMenu.replace('%ID', categoryID));
-    await this.waitForSelectorAndClick(this.categoryMenu.replace('%ID', subCategoryID));
-  }
-
-  /**
-   * Go to login Page
-   * @return {Promise<void>}
-   */
-  async goToLoginPage() {
-    await Promise.all([
-      this.page.waitForNavigation({waitUntil: 'networkidle0'}),
-      this.page.click(this.userInfoLink),
-    ]);
-  }
-
-  /**
-   * Check if customer is connected
-   * @return {Promise<boolean|true>}
-   */
-  async isCustomerConnected() {
-    return this.elementVisible(this.logoutLink, 1000);
   }
 
   /**
